RavenDB-20247 - fix int overflow

+++ src/Sparrow.Server/Meters/TransactionPerformanceMetrics.cs
@@ -88,7 +88,7 @@ namespace Sparrow.Server.Meters
             _tail = null;
 
             var pos = Interlocked.Increment(ref BufferPos);
-            var adjustedTail = pos % Buffer.Length;
+            var adjustedTail = (int)(pos % Buffer.Length);
 
             if (Interlocked.CompareExchange(ref Buffer[adjustedTail], meterItem, null) == null)
                 return;
